backend/requirements.txt
flask==1.1.1
gunicorn==20.0.4
flask_oidc
mysql-connector-python
sqlalchemy
flask_cors
Flask-SQLAlchemy
flask-mongoengine
marshmallow

pymongo